requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.1.0
python-telegram-bot==20.7
langchain==0.1.0
langchain-openai==0.0.5
